learn-agents-v0/pyproject.toml
[project]
name = "learn-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.24.2",
    "decouple>=0.0.7",
    "langchain==1.0.3",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langfuse>=3.5.1",
    "langgraph>=0.6.6",
    "langsmith>=0.4.21",
    "notebook>=7.4.5",
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
